[Hexagon] Delete offload runtime, move files to right places
Within src/runtime/hexagon
- delete directory android,
- move files from hexagon to ., delete hexagon,
- merge host/hexagon_module.cc with hexagon_module.cc, delete host.

Rename HexagonHostModuleNode to HexagonModuleNode.
