+++ src/runtime/hexagon/hexagon_buffer.cc
